+++ common/searchattribute/search_attirbute.go
@@ -27,12 +27,10 @@ package searchattribute
 import (
 	"errors"
 	"fmt"
-	"time"
 
 	commonpb "go.temporal.io/api/common/v1"
 	enumspb "go.temporal.io/api/enums/v1"
 
-	"go.temporal.io/server/common/payload"
 	"go.temporal.io/server/common/service/dynamicconfig"
 )
 
@@ -98,108 +96,6 @@ func GetType(name string, validSearchAttributesFn dynamicconfig.MapPropertyFn) (
 	return getIndexedValueType(saType)
 }
 
-// DecodeValue decodes search attribute value from Payload using (in order):
-// 1. type from MetadataType field,
-// 2. passed type t.
-func DecodeValue(value *commonpb.Payload, t enumspb.IndexedValueType) (interface{}, error) {
-	valueTypeMetadata, metadataHasValueType := value.Metadata[MetadataType]
-	if metadataHasValueType {
-		ivt, err := getIndexedValueType(string(valueTypeMetadata))
-		if err == nil {
-			// MetadataType field has priority over passed type.
-			t = ivt
-		}
-	}
-
-	switch t {
-	case enumspb.INDEXED_VALUE_TYPE_STRING, enumspb.INDEXED_VALUE_TYPE_KEYWORD:
-		var val string
-		if err := payload.Decode(value, &val); err != nil {
-			var listVal []string
-			err = payload.Decode(value, &listVal)
-			return listVal, err
-		}
-		return val, nil
-	case enumspb.INDEXED_VALUE_TYPE_INT:
-		var val int64
-		if err := payload.Decode(value, &val); err != nil {
-			var listVal []int64
-			err = payload.Decode(value, &listVal)
-			return listVal, err
-		}
-		return val, nil
-	case enumspb.INDEXED_VALUE_TYPE_DOUBLE:
-		var val float64
-		if err := payload.Decode(value, &val); err != nil {
-			var listVal []float64
-			err = payload.Decode(value, &listVal)
-			return listVal, err
-		}
-		return val, nil
-	case enumspb.INDEXED_VALUE_TYPE_BOOL:
-		var val bool
-		if err := payload.Decode(value, &val); err != nil {
-			var listVal []bool
-			err = payload.Decode(value, &listVal)
-			return listVal, err
-		}
-		return val, nil
-	case enumspb.INDEXED_VALUE_TYPE_DATETIME:
-		var val time.Time
-		if err := payload.Decode(value, &val); err != nil {
-			var listVal []time.Time
-			err = payload.Decode(value, &listVal)
-			return listVal, err
-		}
-		return val, nil
-	default:
-		return nil, fmt.Errorf("%w: %v", ErrInvalidType, t)
-	}
-}
-
-// EncodeValue encodes search attribute value and IndexedValueType to Payload.
-func EncodeValue(val interface{}, t enumspb.IndexedValueType) (*commonpb.Payload, error) {
-	valPayload, err := payload.Encode(val)
-	if err != nil {
-		return nil, err
-	}
-
-	setMetadataType(valPayload, t)
-	return valPayload, nil
-}
-
-// Encode encodes map of search attribute values to map of Payloads.
-func Encode(searchAttributes map[string]interface{}, validSearchAttributes map[string]enumspb.IndexedValueType) (*commonpb.SearchAttributes, error) {
-	if len(searchAttributes) == 0 {
-		return nil, nil
-	}
-
-	indexedFields := make(map[string]*commonpb.Payload, len(searchAttributes))
-	var lastErr error
-	for saName, saValue := range searchAttributes {
-		valPayload, err := payload.Encode(saValue)
-		if err != nil {
-			lastErr = err
-			indexedFields[saName] = nil
-			continue
-		}
-
-		indexedFields[saName] = valPayload
-
-		if len(validSearchAttributes) == 0 {
-			lastErr = ErrValidMapIsEmpty
-			continue
-		}
-
-		ivt, ok := validSearchAttributes[saName]
-		if !ok {
-			lastErr = fmt.Errorf("%w: %s", ErrInvalidName, saName)
-		}
-		setMetadataType(valPayload, ivt)
-	}
-	return &commonpb.SearchAttributes{IndexedFields: indexedFields}, lastErr
-}
-
 // SetType set type for all valid search attributes which don't have it.
 // It doesn't do any validation and just skip invalid or already set fields.
 func SetType(searchAttributes *commonpb.SearchAttributes, validSearchAttributes map[string]enumspb.IndexedValueType) {
@@ -207,16 +103,16 @@ func SetType(searchAttributes *commonpb.SearchAttributes, validSearchAttributes 
 		return
 	}
 
-	for searchAttributeName, searchAttributePayload := range searchAttributes.GetIndexedFields() {
-		_, metadataHasValueType := searchAttributePayload.Metadata[MetadataType]
+	for saName, saPayload := range searchAttributes.GetIndexedFields() {
+		_, metadataHasValueType := saPayload.Metadata[MetadataType]
 		if metadataHasValueType {
 			continue
 		}
-		valueType, isValidSearchAttribute := validSearchAttributes[searchAttributeName]
+		valueType, isValidSearchAttribute := validSearchAttributes[saName]
 		if !isValidSearchAttribute {
 			continue
 		}
-		setMetadataType(searchAttributePayload, valueType)
+		setMetadataType(saPayload, valueType)
 	}
 }
